% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange-table.express}
\alias{arrange-table.express}
\alias{arrange.ExprBuilder}
\title{Arrange rows}
\usage{
\method{arrange}{ExprBuilder}(.data, ...)
}
\arguments{
\item{.data}{Either a \link[data.table:data.table-class]{data.table::data.table}, or an instance of \link{ExprBuilder}.}

\item{...}{See \link{order_by-table.express}.}
}
\description{
Aliases for \link{order_by-table.express}.
}
