% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-chain.R
\name{chain}
\alias{chain}
\alias{chain.ExprBuilder}
\title{Chain}
\usage{
chain(.data, ...)

\method{chain}{ExprBuilder}(.data, ...)
}
\arguments{
\item{.data}{Object to be chained.}

\item{...}{Arguments for the specific methods.}
}
\description{
Build a chain of similar objects/operations.
}
\details{
The chaining for \link{ExprBuilder} is equivalent to calling \code{\link[=end_expr]{end_expr()}} followed by \code{\link[=start_expr]{start_expr()}}.
}
