% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-EagerExprBuilder.R
\docType{class}
\name{EagerExprBuilder}
\alias{EagerExprBuilder}
\title{Eager frame expression builder}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
EagerExprBuilder
}
\description{
Like \link{ExprBuilder}, but eager in some regards. This shouldn't be used directly.
}
\keyword{datasets}
