% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/end-expr.R
\name{end_expr}
\alias{end_expr}
\title{End and evaluate expression}
\usage{
end_expr(.expr_builder, .parent_env)
}
\arguments{
\item{.expr_builder}{The \link{ExprBuilder} instance.}

\item{.parent_env}{Optionally, the enclosing environment of the expression's environment.}
}
\description{
Helper to evaluate the expression created by \link{ExprBuilder} and get the resulting
\link[data.table:data.table-class]{data.table::data.table}.
}
