% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-expr.R
\name{eval_expr}
\alias{eval_expr}
\alias{eval_expr.ExprBuilder}
\title{Evaluate expression}
\usage{
eval_expr(.data, ...)

\method{eval_expr}{ExprBuilder}(.data, ...)
}
\arguments{
\item{.data}{Something containing an expression that should be evaluated.}

\item{...}{Arguments for the specific methods.}
}
\description{
Helper to evaluate an expression contained in an object to facilitate using
\code{\link[magrittr:pipe]{\%>\%}}.
}
\details{
The method for \link{ExprBuilder} uses \code{\link[base:eval]{base::eval()}}.
}
