% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERBS-filter.R
\name{filter-table.express}
\alias{filter-table.express}
\alias{filter.ExprBuilder}
\title{Filter rows}
\usage{
\method{filter}{ExprBuilder}(.data, ..., .preserve)
}
\arguments{
\item{.data}{An instance of \link{ExprBuilder}.}

\item{...}{See \link{where-table.express}.}

\item{.preserve}{Ignored.}
}
\description{
Alias for \link{where-table.express}.
}
