% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-frame_append.R
\name{frame_append}
\alias{frame_append}
\title{Append expressions to the frame}
\usage{
frame_append(.data, ..., .parse = FALSE)
}
\arguments{
\item{.data}{An \link{ExprBuilder} instance.}

\item{...}{Expressions to add to the frame.}

\item{.parse}{See \link{where-table.express}.}
}
\description{
Add named expressions to the \link[data.table:data.table-class]{data.table::data.table} frame.
}
