% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by.R
\name{group_by-table.express}
\alias{group_by-table.express}
\alias{group_by.ExprBuilder}
\title{Group by clause}
\usage{
\method{group_by}{ExprBuilder}(.data, ..., .parse = FALSE)
}
\arguments{
\item{.data}{Either a \link[data.table:data.table-class]{data.table::data.table}, or an instance of \link{ExprBuilder}.}

\item{...}{Clause for grouping on columns. The \code{by} inside the \code{data.table}'s frame.}

\item{.parse}{See \link{where-table.express}.}
}
\description{
Aggregations for a \link[data.table:data.table-class]{data.table::data.table}.
}
\details{
Everything in \code{...} will be wrapped in a call to \code{list}.
}
