% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-utils.R
\name{is_fun}
\alias{is_fun}
\title{is_fun}
\usage{
is_fun(obj)
}
\arguments{
\item{obj}{Anything really.}
}
\description{
Like \code{\link[rlang:is_function]{rlang::is_function()}} but doesn't throw if the input is maybe something to be quoted.
}
