% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERBS-mutate.R
\name{mutate-table.express}
\alias{mutate-table.express}
\alias{mutate.ExprBuilder}
\title{Add or update columns}
\usage{
\method{mutate}{ExprBuilder}(.data, ..., .parse = FALSE,
  .by_ref = TRUE, .chain = TRUE, .unquote_names = TRUE)
}
\arguments{
\item{.data}{Either a \link[data.table:data.table-class]{data.table::data.table}, or an instance of \link{ExprBuilder}.}

\item{...}{Mutation clauses.}

\item{.parse}{See \link{where-table.express}.}

\item{.by_ref}{Whether to update the data.table by reference or not.}

\item{.chain}{Logical. Should a new frame be automatically chained to the expression if the
clause being set already exists?}

\item{.unquote_names}{Passed to \code{\link[rlang:enexprs]{rlang::enexprs()}}. Set to \code{FALSE} if you want to pass the single
\code{\link[data.table:set]{:=}} expression.}
}
\description{
Add or update columns of a \link[data.table:data.table-class]{data.table::data.table}, possibly by reference using
\code{\link[data.table:set]{:=}}.
}
