% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERBS-mutate_join.R
\name{mutate_join}
\alias{mutate_join}
\title{Mutating join}
\usage{
mutate_join(x, y, ...)
}
\arguments{
\item{x}{In the context of this package, an \link{ExprBuilder} instance.}

\item{y}{In the context of this package, a \link[data.table:data.table-class]{data.table::data.table}.}

\item{...}{Arguments for the specific methods.}
}
\description{
Mutate columns by considering the columns of another object. See \link{joins}.
}
