% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERBS-mutate_sd.R
\name{mutate_sd}
\alias{mutate_sd}
\title{Mutate subset of data}
\usage{
mutate_sd(.data, .how = identity, ..., .SDcols,
  .parse = getOption("table.express.parse", FALSE))
}
\arguments{
\item{.data}{An instance of \link{ExprBuilder}.}

\item{.how}{The transmuting call.}

\item{...}{More arguments for \code{.how} if it is a function or a function-call.}

\item{.SDcols}{See \link[data.table:data.table]{data.table::data.table} and the details here.}

\item{.parse}{Logical. Whether to apply \code{\link[rlang:parse_expr]{rlang::parse_expr()}} to obtain the expressions.}
}
\description{
Like \link{mutate-table.express} but for a single call and some \code{.SDcols}.
}
\details{
This function works similar to \code{\link[=transmute_sd]{transmute_sd()}} but keeps all columns and \emph{can} modify by
reference, like \link{mutate-table.express}. Unlike \code{\link[=transmute_sd]{transmute_sd()}}, \code{.SDcols} \strong{must} be a
character vector.
}
\examples{

data("mtcars")

data.table::as.data.table(mtcars) \%>\%
    start_expr \%>\%
    mutate_sd(.COL * 2, .SDcols = c("mpg", "cyl")) \%>\%
    end_expr

}
