% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERBS-non-equi_joins.R, R/VERBS-rne_join.R
\name{non-equi_joins}
\alias{non-equi_joins}
\alias{rne_join}
\alias{rne_join.ExprBuilder}
\title{Non-equi joins}
\usage{
rne_join(x, y, ...)

\method{rne_join}{ExprBuilder}(x, y, ..., nomatch, mult)
}
\arguments{
\item{x}{An \link{ExprBuilder} instance.}

\item{y}{A \link[data.table:data.table-class]{data.table::data.table}.}

\item{...}{Expressions for the \code{on} part of the join.}

\item{nomatch}{See \link[data.table:data.table]{data.table::data.table}.}

\item{mult}{See \link[data.table:data.table]{data.table::data.table}.}
}
\description{
See also \link[data.table:data.table]{data.table::data.table}.
}
