% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERBS-order_by.R
\name{order_by-table.express}
\alias{order_by-table.express}
\alias{order_by}
\alias{order_by.ExprBuilder}
\title{Order by clause}
\usage{
order_by(.data, ...)

\method{order_by}{ExprBuilder}(.data, ..., .collapse, .parse = FALSE,
  .chain = TRUE)
}
\arguments{
\item{.data}{The input data.}

\item{...}{Arguments for the specific methods.}

\item{.collapse}{Ignored.}

\item{.parse}{See \link{where-table.express}.}

\item{.chain}{Logical. Should a new frame be automatically chained to the expression if the
clause being set already exists?}
}
\description{
Clause for ordering rows.
}
