% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_by.R
\name{order_by}
\alias{order_by}
\title{Order by clause}
\usage{
order_by(.data, ...)
}
\arguments{
\item{.data}{Something resembling a \code{data.frame}.}

\item{...}{Arguments for the specific methods.}
}
\description{
Clause for ordering rows.
}
