% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select-table.express}
\alias{select-table.express}
\alias{select.ExprBuilder}
\alias{select.data.table}
\title{Select clause}
\usage{
\method{select}{ExprBuilder}(.data, ..., with = TRUE)

\method{select}{data.table}(.data, ...)
}
\arguments{
\item{.data}{Either a \link[data.table:data.table-class]{data.table::data.table}, or an instance of \link{ExprBuilder}.}

\item{...}{Clause for selectin/computing on columns. The \code{j} inside the \code{data.table}'s frame. For
the \code{data.table} method, this can also extra parameters from the \code{ExprBuilder} method.}

\item{with}{Passed to \link[data.table:data.table-class]{data.table::data.table}'s \code{\link[data.table:data.table-package]{[}}
method.}
}
\description{
Select or compute over columns of a \link[data.table:data.table-class]{data.table::data.table}.
}
\details{
If the length of \code{...} > 1, everything therein will be wrapped in a call to:
\itemize{
\item \code{list} if \code{with = TRUE}
\item \code{c} otherwise
}
}
