% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-expr.R
\name{start_expr}
\alias{start_expr}
\title{Start expression}
\usage{
start_expr(.DT)
}
\arguments{
\item{.DT}{The data.table.}
}
\description{
Helper to capture a \link[data.table:data.table-class]{data.table::data.table} and begin building an expression for it.
}
