% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{table.express-package}
\alias{table.express}
\alias{table.express-package}
\title{Bindings to use dplyr's verbs with data.table}
\description{
A set of light-weight bindings that map \link[dplyr:dplyr-package]{dplyr::dplyr-package} verbs to an expression that can be
used within a \link[data.table:data.table-package]{data.table::data.table-package}'s frame.
}
\details{
Since this package's functionality is based on the \pkg{rlang} package, and it is still evolving,
breaking changes may be needed in the future.
}
\author{
Alexis Sarda-Espinosa
}
