% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERBS-transmute_sd.R
\name{transmute_sd}
\alias{transmute_sd}
\title{Transmute subset of data}
\usage{
transmute_sd(.data, .fun, ..., .parse = FALSE, .SDcols)
}
\arguments{
\item{.data}{An \link{ExprBuilder} instance.}

\item{.fun}{The function to apply.}

\item{...}{Further arguments for \code{.fun}.}

\item{.parse}{Whether to apply \code{\link[rlang:parse_expr]{rlang::parse_expr()}} to \code{...}.}

\item{.SDcols}{See \link[data.table:data.table]{data.table::data.table}.}
}
\description{
Like \link{transmute-table.express} but for a single function and maybe specifying \code{.SDcols}.
}
\details{
Providing no \code{.SDcols} would work like \code{\link[dplyr:transmute_all]{dplyr::transmute_all()}}.
}
