% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{where-table.express}
\alias{where-table.express}
\alias{where.ExprBuilder}
\title{Where clause}
\usage{
\method{where}{ExprBuilder}(.data, ..., .collapse = `&`,
  .parse = FALSE)
}
\arguments{
\item{.data}{Either a \link[data.table:data.table-class]{data.table::data.table}, or an instance of \link{ExprBuilder}.}

\item{...}{Clause for subsetting. The \code{i} inside the \code{data.table}'s frame. For the \code{data.table}
method, this can also extra parameters from the \code{ExprBuilder} method.}

\item{.collapse}{A boolean function which will be used to "concatenate" all conditions in \code{...}.}

\item{.parse}{If you want/need to provide strings in \code{...}, set this to \code{TRUE} to call
\code{\link[rlang:parse_expr]{rlang::parse_expr()}} on each one.}
}
\description{
Clause for subsetting rows of a \link[data.table:data.table-class]{data.table::data.table}.
}
