% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where.R
\name{where}
\alias{where}
\title{Where clause}
\usage{
where(.data, ...)
}
\arguments{
\item{.data}{Something resembling a \code{data.frame}.}

\item{...}{Arguments for the specific methods.}
}
\description{
Clause for subsetting rows.
}
