% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERBS-arrange.R
\name{arrange-table.express}
\alias{arrange-table.express}
\alias{arrange.ExprBuilder}
\title{Arrange rows}
\usage{
\method{arrange}{ExprBuilder}(.data, ...)
}
\arguments{
\item{.data}{An instance of \link{ExprBuilder}.}

\item{...}{See \link{order_by-table.express}.}
}
\description{
Alias for \link{order_by-table.express}.
}
