% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERBS-arrange.R
\name{arrange-table.express}
\alias{arrange-table.express}
\alias{arrange.ExprBuilder}
\alias{arrange.data.table}
\title{Arrange rows}
\usage{
\method{arrange}{ExprBuilder}(.data, ...)

\method{arrange}{data.table}(.data, ...)
}
\arguments{
\item{.data}{An instance of \link{ExprBuilder}.}

\item{...}{See \link{order_by-table.express}.}
}
\description{
Alias for \link{order_by-table.express}.
}
\details{
To see more examples, check the
\href{https://asardaes.github.io/table.express/articles/table.express.html}{vignette}, or the
\link{table.express-package} entry.
}
