% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{chain}
\alias{chain}
\alias{chain.ExprBuilder}
\title{Chain anything}
\usage{
chain(.data, ...)

\method{chain}{ExprBuilder}(.data, ...)
}
\arguments{
\item{.data}{Object to be chained.}

\item{...}{Arguments for the specific methods.}
}
\description{
Build a chain of similar objects/operations.
}
\details{
The \link{ExprBuilder} method is equivalent to calling \code{\link[=end_expr]{end_expr()}} and then \code{\link[=start_expr]{start_expr()}} again.
}
