% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DELIMITERS-chain.R
\name{chain}
\alias{chain}
\alias{chain.ExprBuilder}
\title{Chain}
\usage{
chain(.data, ...)

\method{chain}{ExprBuilder}(.data, ..., .parent_env = rlang::caller_env())
}
\arguments{
\item{.data}{Object to be chained.}

\item{...}{Arguments for the specific methods.}

\item{.parent_env}{See \code{\link[=end_expr]{end_expr()}}.}
}
\description{
Build a chain of similar objects/operations.
}
\details{
The chaining for \link{ExprBuilder} is equivalent to calling \code{\link[=end_expr]{end_expr()}} followed by \code{\link[=start_expr]{start_expr()}}.
The ellipsis (\code{...}) is passed to both functions.

To see more examples, check the
\href{https://asardaes.github.io/table.express/articles/table.express.html}{vignette}, or the
\link{table.express-package} entry.
}
