% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERBS-distinct.R
\name{distinct-table.express}
\alias{distinct-table.express}
\alias{distinct.ExprBuilder}
\title{Rows with distinct combinations of columns}
\usage{
\method{distinct}{ExprBuilder}(.data, ..., .keep_all = TRUE,
  .keep_old = FALSE, .parse = getOption("table.express.parse", FALSE))
}
\arguments{
\item{.data}{An instance of \link{ExprBuilder}.}

\item{...}{Which columns to use to determine uniqueness.}

\item{.keep_all}{Whether to also keep columns that are not mentioned in \code{...}.}

\item{.keep_old}{This should be changed to \code{TRUE} if a new column is created in \code{...} and the
columns used to create it should also be kept. Has no effect if \code{.keep_all} is \code{FALSE}. See
examples.}

\item{.parse}{Logical. Whether to apply \code{\link[rlang:parse_expr]{rlang::parse_expr()}} to obtain the expressions.}
}
\description{
Rows with distinct combinations of columns
}
\details{
To see more examples, check the
\href{https://asardaes.github.io/table.express/articles/table.express.html}{vignette}, or the
\link{table.express-package} entry.
}
\examples{

data("mtcars")
data.table::as.data.table(mtcars) \%>\%
    start_expr \%>\%
    distinct(amvs = am + vs, .keep_old = TRUE) \%>\%
    end_expr

}
