% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERBS-distinct.R
\name{distinct-table.express}
\alias{distinct-table.express}
\alias{distinct.ExprBuilder}
\alias{distinct.data.table}
\title{Rows with distinct combinations of columns}
\usage{
\method{distinct}{ExprBuilder}(.data, ..., .keep = TRUE,
  .parse = getOption("table.express.parse", FALSE))

\method{distinct}{data.table}(.data, ...)
}
\arguments{
\item{.data}{An instance of \link{ExprBuilder}.}

\item{...}{Which columns to use to determine uniqueness.}

\item{.keep}{See details below.}

\item{.parse}{Logical. Whether to apply \code{\link[rlang:parse_expr]{rlang::parse_expr()}} to obtain the expressions.}
}
\description{
Rows with distinct combinations of columns
}
\details{
If \code{.keep = TRUE} (the default), the columns not mentioned in \code{...} are also kept. However, if
a new column is created in one of the expressions therein, \code{.keep} can also be set to a character
vector containing the names of \emph{all} the columns that should be in the result in addition to the
ones mentioned in \code{...}. See the examples.

To see more examples, check the
\href{https://asardaes.github.io/table.express/articles/table.express.html}{vignette}, or the
\link{table.express-package} entry.
}
\examples{

data("mtcars")

# compare with .keep = TRUE
data.table::as.data.table(mtcars) \%>\%
    start_expr \%>\%
    distinct(amvs = am + vs, .keep = names(mtcars)) \%>\%
    end_expr

}
