% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-EagerExprBuilder.R
\docType{class}
\name{EagerExprBuilder}
\alias{EagerExprBuilder}
\title{Eager frame expression builder}
\description{
Like \link{ExprBuilder}, but eager in some regards. This shouldn't be used directly.
}
\section{Super class}{
\code{\link[table.express:ExprBuilder]{table.express::ExprBuilder}} -> \code{EagerExprBuilder}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{EagerExprBuilder$new()}}
\item \href{#method-chain}{\code{EagerExprBuilder$chain()}}
\item \href{#method-chain_if_set}{\code{EagerExprBuilder$chain_if_set()}}
\item \href{#method-clone}{\code{EagerExprBuilder$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="table.express" data-topic="ExprBuilder" data-id="eval">}\href{../../table.express/html/ExprBuilder.html#method-eval}{\code{table.express::ExprBuilder$eval()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="table.express" data-topic="ExprBuilder" data-id="print">}\href{../../table.express/html/ExprBuilder.html#method-print}{\code{table.express::ExprBuilder$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="table.express" data-topic="ExprBuilder" data-id="seek_and_nestroy">}\href{../../table.express/html/ExprBuilder.html#method-seek_and_nestroy}{\code{table.express::ExprBuilder$seek_and_nestroy()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="table.express" data-topic="ExprBuilder" data-id="set_by">}\href{../../table.express/html/ExprBuilder.html#method-set_by}{\code{table.express::ExprBuilder$set_by()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="table.express" data-topic="ExprBuilder" data-id="set_i">}\href{../../table.express/html/ExprBuilder.html#method-set_i}{\code{table.express::ExprBuilder$set_i()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="table.express" data-topic="ExprBuilder" data-id="set_j">}\href{../../table.express/html/ExprBuilder.html#method-set_j}{\code{table.express::ExprBuilder$set_j()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="table.express" data-topic="ExprBuilder" data-id="tidy_select">}\href{../../table.express/html/ExprBuilder.html#method-tidy_select}{\code{table.express::ExprBuilder$tidy_select()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EagerExprBuilder$new(DT)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{DT}}{A \link[data.table:data.table-class]{data.table::data.table}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chain"></a>}}
\if{latex}{\out{\hypertarget{method-chain}{}}}
\subsection{Method \code{chain()}}{
Override to abort if chaining is attempted.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EagerExprBuilder$chain(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-chain_if_set"></a>}}
\if{latex}{\out{\hypertarget{method-chain_if_set}{}}}
\subsection{Method \code{chain_if_set()}}{
Override to abort if chaining is attempted.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EagerExprBuilder$chain_if_set(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Ignored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EagerExprBuilder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
