% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-ExprBuilder.R
\docType{class}
\name{ExprBuilder}
\alias{ExprBuilder}
\title{Frame expression builder}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ExprBuilder
}
\description{
Build an expression that will be used inside a \link[data.table:data.table-class]{data.table::data.table}'s frame.
}
\section{Fields}{

\describe{
\item{\code{where}}{Clause for subsetting. The \code{i} inside the \code{data.table}'s frame.}

\item{\code{select}}{Clause for selectin/computing on columns. The \code{j} inside the \code{data.table}'s frame.}

\item{\code{expr}}{The final expression that can be evaluated with \code{\link[base:eval]{base::eval()}} or
\code{\link[rlang:eval_bare]{rlang::eval_bare()}}.}
}}

\section{Methods}{


\describe{
\item{\code{initialize(DT)}}{Constructor that receives a \link[data.table:data.table-class]{data.table::data.table} in \code{DT}.}
\item{\code{chain()}}{Start a new expression with the current one as its parent.}
\item{\code{eval(env)}}{Evaluate the final expression with \code{env} as the enclosing environment.}
\item{\code{print(...)}}{Prints the built \code{expr}.}
}
}

\keyword{datasets}
