% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter-table.express}
\alias{filter-table.express}
\alias{filter.ExprBuilder}
\title{Filter rows}
\usage{
\method{filter}{ExprBuilder}(.data, ..., .preserve)
}
\arguments{
\item{.data}{Either a \link[data.table:data.table-class]{data.table::data.table}, or an instance of \link{ExprBuilder}.}

\item{...}{See \link{where-table.express}.}

\item{.preserve}{Ignored.}
}
\description{
Aliases for \link{where-table.express}.
}
