% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERBS-joins.R
\name{joins}
\alias{joins}
\alias{joins-table.express}
\alias{left_join.ExprBuilder}
\title{Joining verbs}
\usage{
\method{left_join}{ExprBuilder}(x, y, ..., .adding)
}
\arguments{
\item{x}{An \link{ExprBuilder} instance.}

\item{y}{A \link[data.table:data.table-class]{data.table::data.table}.}

\item{...}{Expressions for the \code{on} part of the join.}

\item{.adding}{A possibly named character vector specifying which columns from \code{y} are part of
the join.}
}
\description{
See also \link[dplyr:join]{dplyr::join}.
}
\details{
TBD
}
\examples{

lhs <- data.table::data.table(x = rep(c("b", "a", "c"), each = 3),
                              y = c(1, 3, 6),
                              v = 1:9)

rhs <- data.table::data.table(x = c("c", "b"),
                              v = 8:7,
                              foo = c(4, 2))


# creates new data.table eagerly
lhs \%>\%
    start_expr \%>\%
    left_join(rhs, "x") \%>\%
    end_expr

# would modify lhs by reference lazily
lhs \%>\%
    start_expr \%>\%
    left_join(rhs, "x", .adding = c("foo", i.v = "v"))

}
