% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate-table.express}
\alias{mutate-table.express}
\alias{mutate.ExprBuilder}
\title{Add or update columns}
\usage{
\method{mutate}{ExprBuilder}(.data, ..., .parse = FALSE,
  .by_ref = TRUE)
}
\arguments{
\item{.data}{Either a \link[data.table:data.table-class]{data.table::data.table}, or an instance of \link{ExprBuilder}.}

\item{...}{Mutation clauses.}

\item{.parse}{See \link{where-table.express}.}

\item{.by_ref}{Whether to update the data.table by reference or not.}
}
\description{
Add or update columns of a \link[data.table:data.table-class]{data.table::data.table}, possibly by reference using
\code{\link[data.table:set]{:=}}.
}
