% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERBS-non-equi_joins.R, R/VERBS-lne_join.R,
%   R/VERBS-rne_join.R
\name{non-equi_joins}
\alias{non-equi_joins}
\alias{lne_join}
\alias{lne_join.ExprBuilder}
\alias{rne_join}
\alias{rne_join.ExprBuilder}
\title{Non-equi joins}
\usage{
lne_join(x, y, ..., nomatch, mult)

\method{lne_join}{ExprBuilder}(x, y, ..., nomatch, mult)

rne_join(x, y, ..., nomatch, mult)

\method{rne_join}{ExprBuilder}(x, y, ..., nomatch, mult)
}
\arguments{
\item{x}{An \link{ExprBuilder} instance.}

\item{y}{A \link[data.table:data.table-class]{data.table::data.table}.}

\item{...}{Expressions for the \code{on} part of the join.}

\item{nomatch}{See \link[data.table:data.table]{data.table::data.table}.}

\item{mult}{See \link[data.table:data.table]{data.table::data.table}.}
}
\description{
Left and right non-equi joins.
}
\details{
Non-equi joins are similar to normal joins but support comparisons (\code{<}, \code{>}, \code{<=}, and \code{>=}) in
the \code{on} expression. We can loosely divide them into left and right non-equi joins, \code{lne_join}
and \code{rne_join} respectively. Their differences are the rows in the result, and the values
returned in a given column that is part of the comparison. Check the "Joining verbs" vignette for
more information.
}
\seealso{
\link{joins}, \link[data.table:data.table]{data.table::data.table}, \link[dplyr:join]{dplyr::join}
}
