% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_by.R
\name{order_by-table.express}
\alias{order_by-table.express}
\alias{order_by}
\alias{order_by.ExprBuilder}
\title{Order by clause}
\usage{
order_by(.data, ...)

\method{order_by}{ExprBuilder}(.data, ..., .collapse, .parse = FALSE)
}
\arguments{
\item{.data}{The input data.}

\item{...}{Arguments for the specific methods.}

\item{.collapse}{Ignored.}

\item{.parse}{See \link{where-table.express}.}
}
\description{
Clause for ordering rows.
}
