% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_by.R
\name{order_by-table.express}
\alias{order_by-table.express}
\alias{order_by.ExprBuilder}
\title{Order by clause}
\usage{
\method{order_by}{ExprBuilder}(.data, ..., .collapse, .parse = FALSE)
}
\arguments{
\item{.data}{Either a \link[data.table:data.table-class]{data.table::data.table}, or an instance of \link{ExprBuilder}.}

\item{...}{The columns to sort by.}

\item{.collapse}{Ignored.}

\item{.parse}{See \link{where-table.express}.}
}
\description{
Clause for ordering rows of a \link[data.table:data.table-class]{data.table::data.table}.
}
