% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R, R/filter.R, R/group_by.R,
%   R/mutate.R, R/select.R, R/transmute.R, R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{arrange}
\alias{filter}
\alias{group_by}
\alias{mutate}
\alias{select}
\alias{transmute}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{transmute}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

