% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-utils.R, R/VERBS-arrange.R,
%   R/VERBS-filter.R, R/VERBS-group_by.R, R/VERBS-mutate.R, R/VERBS-select.R,
%   R/VERBS-transmute.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{arrange}
\alias{filter}
\alias{group_by}
\alias{mutate}
\alias{select}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{everything}
\alias{last_col}
\alias{transmute}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{transmute}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{tidyselect}{\code{\link[tidyselect]{starts_with}}, \code{\link[tidyselect]{ends_with}}, \code{\link[tidyselect]{contains}}, \code{\link[tidyselect]{matches}}, \code{\link[tidyselect]{num_range}}, \code{\link[tidyselect]{one_of}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect]{last_col}}}
}}

