% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-tidyselect.R, R/VERBS-anti_join.R,
%   R/VERBS-arrange.R, R/VERBS-distinct.R, R/VERBS-filter.R,
%   R/VERBS-full_join.R, R/VERBS-group_by.R, R/VERBS-inner_join.R,
%   R/VERBS-left_join.R, R/VERBS-mutate.R, R/VERBS-right_join.R,
%   R/VERBS-select.R, R/VERBS-semi_join.R, R/VERBS-transmute.R, R/pkg.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{everything}
\alias{last_col}
\alias{anti_join}
\alias{arrange}
\alias{distinct}
\alias{filter}
\alias{full_join}
\alias{group_by}
\alias{inner_join}
\alias{left_join}
\alias{mutate}
\alias{right_join}
\alias{select}
\alias{semi_join}
\alias{transmute}
\alias{:=}
\alias{\%>\%}
\alias{\%T>\%}
\alias{!!}
\alias{!!!}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table]{:=}}}

  \item{dplyr}{\code{\link[dplyr]{anti_join}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{full_join}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{inner_join}}, \code{\link[dplyr]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{right_join}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{semi_join}}, \code{\link[dplyr]{transmute}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%T>\%}}}

  \item{rlang}{\code{\link[rlang]{!!}}, \code{\link[rlang]{!!!}}}

  \item{tidyselect}{\code{\link[tidyselect]{starts_with}}, \code{\link[tidyselect]{ends_with}}, \code{\link[tidyselect]{contains}}, \code{\link[tidyselect]{matches}}, \code{\link[tidyselect]{num_range}}, \code{\link[tidyselect]{one_of}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect]{last_col}}}
}}

