% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTILS-tidyselect.R, R/VERBS-anti_join.R,
%   R/VERBS-arrange.R, R/VERBS-distinct.R, R/VERBS-filter.R,
%   R/VERBS-full_join.R, R/VERBS-group_by.R, R/VERBS-inner_join.R,
%   R/VERBS-left_join.R, R/pkg.R, R/VERBS-mutate.R, R/VERBS-right_join.R,
%   R/VERBS-select.R, R/VERBS-semi_join.R, R/VERBS-summarize.R,
%   R/VERBS-transmute.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{starts_with}
\alias{ends_with}
\alias{contains}
\alias{matches}
\alias{num_range}
\alias{one_of}
\alias{everything}
\alias{last_col}
\alias{anti_join}
\alias{arrange}
\alias{distinct}
\alias{filter}
\alias{full_join}
\alias{group_by}
\alias{inner_join}
\alias{left_join}
\alias{:=}
\alias{\%>\%}
\alias{\%T>\%}
\alias{!!}
\alias{!!!}
\alias{mutate}
\alias{right_join}
\alias{select}
\alias{semi_join}
\alias{summarize}
\alias{summarise}
\alias{transmute}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table:assign]{:=}}}

  \item{dplyr}{\code{\link[dplyr:filter-joins]{anti_join}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr:mutate-joins]{full_join}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr:mutate-joins]{inner_join}}, \code{\link[dplyr:mutate-joins]{left_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr:mutate-joins]{right_join}}, \code{\link[dplyr]{select}}, \code{\link[dplyr:filter-joins]{semi_join}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr:summarise]{summarize}}, \code{\link[dplyr]{transmute}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}, \code{\link[magrittr:tee]{\%T>\%}}}

  \item{rlang}{\code{\link[rlang:injection-operator]{!!}}, \code{\link[rlang:splice-operator]{!!!}}}

  \item{tidyselect}{\code{\link[tidyselect:starts_with]{contains}}, \code{\link[tidyselect:starts_with]{ends_with}}, \code{\link[tidyselect]{everything}}, \code{\link[tidyselect:everything]{last_col}}, \code{\link[tidyselect:starts_with]{matches}}, \code{\link[tidyselect:starts_with]{num_range}}, \code{\link[tidyselect]{one_of}}, \code{\link[tidyselect]{starts_with}}}
}}

