% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DELIMITERS-start_expr.R
\name{start_expr}
\alias{start_expr}
\alias{start_expr.data.table}
\title{Start expression}
\usage{
start_expr(.data, ...)

\method{start_expr}{data.table}(.data, ...,
  .verbose = getOption("table.express.verbose", FALSE))
}
\arguments{
\item{.data}{Optionally, something to capture for the expression.}

\item{...}{Arguments for the specific methods.}

\item{.verbose}{Whether to print more information during the expression-building process.}
}
\description{
Start building an expression.
}
\details{
The \link[data.table:data.table-class]{data.table::data.table} method returns an \link{ExprBuilder} instance.

To see more examples, check the
\href{https://asardaes.github.io/table.express/articles/table.express.html}{vignette}, or the
\link{table.express-package} entry.
}
