% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERBS-transmute.R
\name{transmute-table.express}
\alias{transmute-table.express}
\alias{transmute.ExprBuilder}
\title{Add or adjust columns and keep only the result}
\usage{
\method{transmute}{ExprBuilder}(.data, ..., .parse = FALSE,
  .chain = TRUE)
}
\arguments{
\item{.data}{An instance of \link{ExprBuilder}.}

\item{...}{Transmutation clauses.}

\item{.parse}{Logical. Whether to apply \code{\link[rlang:parse_expr]{rlang::parse_expr()}} to obtain the expressions.}

\item{.chain}{Logical. Should a new frame be automatically chained to the expression if the
clause being set already exists?}
}
\description{
Add or adjust columns of a \link[data.table:data.table-class]{data.table::data.table}, keeping only the resulting columns.
}
\details{
This first calls \link{mutate-table.express} and chains a \link{select-table.express} operation to get only
the resulting columns.
}
