% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VERBS-where.R
\name{where-table.express}
\alias{where-table.express}
\alias{where}
\alias{where.ExprBuilder}
\title{Where clause}
\usage{
where(.data, ...)

\method{where}{ExprBuilder}(.data, ..., .collapse = `&`,
  .parse = FALSE, .chain = TRUE)
}
\arguments{
\item{.data}{The input data.}

\item{...}{Arguments for the specific methods.}

\item{.collapse}{A boolean function which will be used to "concatenate" all conditions in \code{...}.}

\item{.parse}{If you want/need to provide strings in \code{...}, set this to \code{TRUE} to call
\code{\link[rlang:parse_expr]{rlang::parse_expr()}} on each one.}

\item{.chain}{Logical. Should a new frame be automatically chained to the expression if the
clause being set already exists?}
}
\description{
Clause for subsetting rows.
}
